+++ src/layouts/Breadcrumbs.jsx
@@ -1,0 +1,53 @@
+// React Imports
+import React from 'react';
+import { Link, useLocation } from 'react-router-dom';
+// Material UI Imports
+import { Breadcrumbs as MUIBreadcrumbs } from '@mui/material';
+import NavigateNextIcon from '@mui/icons-material/NavigateNext';
+import Typography from '@mui/material/Typography';
+
+const Breadcrumbs = () => {
+  const location = useLocation();
+
+  const crumbs = location?.pathname
+    .split('/')
+    .slice(1)
+    .map((crumb) => {
+      let string = crumb;
+      string = decodeURIComponent(crumb.replace('-', ' '));
+      if (!string.includes('http')) {
+        string = string
+          .split(' ')
+          .map((s) => s.charAt(0).toUpperCase() + s.slice(1))
+          .join(' ');
+      }
+      return string;
+    });
+
+  return (
+    <MUIBreadcrumbs
+      aria-label="breadcrumb"
+      separator={<NavigateNextIcon fontSize="small" />}
+      sx={{ margin: '20px 80px', width: '100%' }}
+    >
+      {crumbs.map((crumb, index) =>
+        index !== crumbs.length - 1 ? (
+          <Link
+            to={`/${crumbs
+              .slice(0, index + 1)
+              .join('/')
+              .replaceAll(' ', '-')
+              .toLowerCase()}`}
+            key={crumb}
+          >
+            {crumb}
+          </Link>
+        ) : (
+          <Typography key={crumb}>{crumb}</Typography>
+        )
+      )}
+    </MUIBreadcrumbs>
+  );
+};
+
+export default Breadcrumbs;
